Restructuring code tools into local python packages
