+++ mre_code_tools/pipeline/compile/mre_generate_councilor_editor_scripted_triggers.py
