+++ mre_code_tools/pipeline/compile/mre_stitch_gui_files.py
