+++ mre_code_tools/pipeline/extract/stellaris_yaml_converter.py
